requirements.txt
autorank==1.3.0
grad_cam==1.5.5
kagglehub==0.3.12
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.1
Pillow==9.0.1
Pillow==11.3.0
scikit_learn==1.7.1
torch==2.7.1
torchmetrics==1.8.1
torchvision==0.22.1
tqdm==4.67.1
